+++ stellar-dotnet-sdk/requests/OrderBookRequestBuilder.cs
@@ -37,19 +37,16 @@ namespace stellar_dotnet_sdk.requests
             return this;
         }
 
-
         public override OrderBookRequestBuilder Cursor(string token)
         {
             throw new NotImplementedException();
         }
 
-
         public override OrderBookRequestBuilder Limit(int number)
         {
             throw new NotImplementedException();
         }
-
-
+        
         public override OrderBookRequestBuilder Order(OrderDirection direction)
         {
             throw new NotImplementedException();
