+++ stellar-dotnet-sdk/responses/AssetResponse.cs
@@ -81,30 +81,6 @@ namespace stellar_dotnet_sdk.responses
     /// <summary>
     /// 
     /// </summary>
-    public class AssetResponseFlags
-    {
-        /// <summary>
-        /// 
-        /// </summary>
-        [JsonProperty(PropertyName = "auth_required")]
-        public bool AuthRequired { get; set; }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        [JsonProperty(PropertyName = "auth_revocable")]
-        public bool AuthRevocable { get; set; }
-
-        public AssetResponseFlags(bool authRequired, bool authRevocable)
-        {
-            AuthRequired = authRequired;
-            AuthRevocable = authRevocable;
-        }
-    }
-
-    /// <summary>
-    /// 
-    /// </summary>
     public class AssetResponseLinks
     {
         /// <summary>
