+++ stellar-dotnet-sdk/responses/FriendBotResponseLinks.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace stellar_dotnet_sdk.responses
+{
+    public class FriendBotResponseLinks
+    {
+        [JsonProperty(PropertyName = "transaction")]
+        public Link Transaction { get; private set; }
+    }
+}
