+++ stellar-dotnet-sdk/responses/OfferResponse.cs
@@ -45,18 +45,4 @@ namespace stellar_dotnet_sdk.responses
         [JsonProperty(PropertyName = "_links")]
         public OfferResponseLinks Links { get; private set; }
     }
-
-    public class OfferResponseLinks
-    {
-        public OfferResponseLinks(Link self, Link offerMager)
-        {
-            Self = self;
-            OfferMager = offerMager;
-        }
-
-        [JsonProperty(PropertyName = "self")] public Link Self { get; private set; }
-
-        [JsonProperty(PropertyName = "offer_maker")]
-        public Link OfferMager { get; private set; }
-    }
 }
