+++ stellar-dotnet-sdk/responses/PathResponse.cs
@@ -52,14 +52,4 @@ namespace stellar_dotnet_sdk.responses
             Links = links;
         }
     }
-
-    public class PathResponseLinks
-    {
-        [JsonProperty(PropertyName = "self")] public Link Self { get; }
-
-        public PathResponseLinks(Link self)
-        {
-            Self = self;
-        }
-    }
 }
