+++ stellar-dotnet-sdk/responses/PathResponseLinks.cs
@@ -1,0 +1,14 @@
+﻿using Newtonsoft.Json;
+
+namespace stellar_dotnet_sdk.responses
+{
+    public class PathResponseLinks
+    {
+        [JsonProperty(PropertyName = "self")] public Link Self { get; }
+
+        public PathResponseLinks(Link self)
+        {
+            Self = self;
+        }
+    }
+}
