+++ stellar-dotnet-sdk/responses/SubmitTransactionResponse.cs
@@ -7,9 +7,8 @@ namespace stellar_dotnet_sdk.responses
 {
     public class SubmitTransactionResponse : Response
     {
-        private readonly string _EnvelopeXdr;
-        private readonly string _ResultXdr;
-
+        private readonly string _envelopeXdr;
+        private readonly string _resultXdr;
 
         [JsonProperty(PropertyName = "hash")] public string Hash { get; private set; }
 
@@ -23,7 +22,7 @@ namespace stellar_dotnet_sdk.responses
             {
                 if (IsSuccess())
                 {
-                    return _EnvelopeXdr;
+                    return _envelopeXdr;
                 }
                 else
                 {
@@ -39,7 +38,7 @@ namespace stellar_dotnet_sdk.responses
             {
                 if (IsSuccess())
                 {
-                    return _ResultXdr;
+                    return _resultXdr;
                 }
                 else
                 {
@@ -56,8 +55,8 @@ namespace stellar_dotnet_sdk.responses
             SubmitTransactionResponseExtras = extras;
             Ledger = ledger;
             Hash = hash;
-            _EnvelopeXdr = envelopeXdr;
-            _ResultXdr = resultXdr;
+            _envelopeXdr = envelopeXdr;
+            _resultXdr = resultXdr;
         }
 
         public bool IsSuccess()
@@ -144,8 +143,8 @@ namespace stellar_dotnet_sdk.responses
 
                 public ResultCodes(string transactionResultCode, List<string> operationsResultCodes)
                 {
-                    this.TransactionResultCode = transactionResultCode;
-                    this.OperationsResultCodes = operationsResultCodes;
+                    TransactionResultCode = transactionResultCode;
+                    OperationsResultCodes = operationsResultCodes;
                 }
             }
         }
