+++ stellar-dotnet-sdk/responses/effects/TradeEffectResponse.cs
@@ -32,7 +32,6 @@ namespace stellar_dotnet_sdk.responses.effects
         [JsonProperty(PropertyName = "offer_id")]
         public long OfferId { get; }
 
-
         [JsonProperty(PropertyName = "sold_amount")]
         public string SoldAmount { get; }
